# Lazily allocate TypedArena's first chunk