@nnethercote 2016-09-20
I replaced the `len() > 0` expressions in the original version with `is_empty()`.