@bluss 2016-09-20
In the libcollections convention, `with_capacity` is for explicit up front allocation of capacity, while new is welcome to not allocate anything. Since the compiler exclusively uses `new` what I can see, is it not best to follow the convention here?

Alternatively, `with_capacity` is already a bit outside the convention since it's really, `with_chunk_size` or something like that, so it can be renamed.

Either way, doc just needs update to not say "preallocated" for new and with_capacity.