@bluss 2016-09-20
The is_empty tests seem to be inverted