@nnethercote 2016-09-20
Thank you for the comments, @bluss. I fixed the inverted is_empty tests and updated the comments for `new` and `with_capacity`.