@TimNN 2016-09-20
I did a quick grep over the rust source code and I don't think `TypedArena::with_capacity` is ever used, so it *may* be possible to just remove it entirely.