@nnethercote 2016-09-20
I made the requested control flow changes. I haven't changed `with_capacity`, though I'm happy to remove it if there is consensus there.