@nnethercote 2016-09-20
Note that `clear` and `drop` are now very similar, although one uses `drain` and the other `iter_mut`. I don't know if that similarity can be factored out.