@arielb1 2016-09-20
Nice catch @nnethercote! The redundant arenas used to not matter because we had 1 `CtxtArenas` struct per compiler run, but we missed the overhead when we moved to 1 arena/function.