@nnethercote 2016-09-20
Now that I have a better idea of how rustc-benchmarks works, here are some 
updated numbers. This is with a rustc configured with '--enable-optimize
--enable-debuginfo', producing debug builds.

stage 1 (uses glibc malloc)
```
futures-rs-test-all                  4.925s vs  4.755s --> 1.036x faster
helloworld                           0.220s vs  0.221s --> 0.995x faster
html5ever-2016-08-25                23.086s vs 22.216s --> 1.039x faster
hyper.0.5.0                         21.441s vs 20.491s --> 1.046x faster
inflate-0.1.0                        5.083s vs  4.860s --> 1.046x faster
issue-32062-equality-relations-c...  0.397s vs  0.396s --> 1.003x faster
issue-32278-big-array-of-strings     1.839s vs  1.837s --> 1.001x faster
jld-day15-parser                     5.805s vs  5.656s --> 1.026x faster
piston-image-0.10.3                 28.530s vs 27.061s --> 1.054x faster
regex.0.1.30                         2.975s vs  2.798s --> 1.063x faster
rust-encoding-0.3.0                  3.571s vs  3.537s --> 1.010x faster
syntex-0.42.2                       52.195s vs 49.760s --> 1.049x faster
syntex-0.42.2-incr-clean            52.023s vs 49.806s --> 1.045x faster
```
stage2 (uses jemalloc)
```
futures-rs-test-all                  4.283s vs  4.188s --> 1.023x faster
helloworld                           0.222s vs  0.221s --> 1.005x faster
html5ever-2016-08-25                17.508s vs 17.154s --> 1.021x faster
hyper.0.5.0                         17.506s vs 17.164s --> 1.020x faster
inflate-0.1.0                        4.410s vs  4.380s --> 1.007x faster
issue-32062-equality-relations-c...  0.366s vs  0.362s --> 1.011x faster
issue-32278-big-array-of-strings     1.636s vs  1.650s --> 0.992x faster
jld-day15-parser                     4.698s vs  4.646s --> 1.011x faster
piston-image-0.10.3                 23.283s vs 22.819s --> 1.020x faster
regex.0.1.30                         2.527s vs  2.460s --> 1.027x faster
rust-encoding-0.3.0                  3.279s vs  3.315s --> 0.989x faster
syntex-0.42.2                       42.986s vs 42.215s --> 1.018x faster
syntex-0.42.2-incr-clean            43.079s vs 42.134s --> 1.022x faster
```
With glibc malloc they're mostly in the range 1.03--1.06x faster. With jemalloc 
they're mostly in the range 1.01--1.03x faster. The couple that look slower are
just due to measurement noise.