@eddyb 2016-09-21
@nnethercote FWIW I've been meaning to eventually move to a single common drop-less arena (instead of a dozen typed ones), but there were things to rework first to make that even possible - we're almost there, in fact `Ty` only has `TraitObject` left that's not POD (I think I want that to be a slice of existential predicates), and everything else has at most a `Vec` somewhere (which can become an arena slice).